app/django-bbx/requirements.txt
Django==1.5.1
South==0.8.1
argparse==1.2.1
django-extensions==1.1.1
djangorestframework==2.3.6
six==1.3.0
wsgiref==0.1.2
